Combined test data initialisers
There is a dependency from the Protected Species data on the general Test Data and this change makes that clear.
